removed design smell by extracting new class URLClassUtil from DemoChartsUtil
